requirements.txt
scikit-learn==0.21.3
rasterio==1.0.21
geopandas==0.4.1
segmentation-models-pytorch==0.0.3
catalyst[dl]==19.10
loguru==0.3.2
wandb==0.8.12
torch==1.2.0
safitty==1.2.5

